backend/requirements.txt
openai
python-dotenv
speechrecognition
pyaudio
pyttsx3
requests
pydantic
